Add upload file feature
Added include to the new upload.php file that handles file upload.
The file upload.php uses the function uploadFiles that receives three 3
parameters that are name, tmp_name,  and size to upload the file.
This parameters come from  a single form input of type-“file”.   Only
when the function uploadFile returns true, the changes will be added to
the database.